requirements.txt
Cython==3.0.12
GitPython==3.1.44
MarkupSafe==3.0.2
PyYAML==6.0.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dtw-python==1.5.3
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.0
gitdb==4.0.12
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.44.0
more-itertools==10.6.0
mpmath==1.3.0
narwhals==1.31.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
sacremoses==0.1.1
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
setuptools==77.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.0
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wheel==0.45.1
whisper-timestamped==1.15.8
autocommand==2.2.2
backports.tarfile==1.2.0
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
platformdirs==4.2.2
tomli==2.0.1
typeguard==4.3.0
zipp==3.19.2
